requirements.txt
fastapi # Ensure you are using the latest version compatible with your setup
python
pydantic
uvicorn
pymongo
bcrypt
python-dotenv
pyjwt
